+++ src/BrainGraph.ComputeRT/SingleDatatypeSupport.h
@@ -14,13 +14,13 @@ namespace BrainGraph { namespace Compute { namespace Graph
 			dv1 = 0, dv2 = 0;
 		}
 
-		void IncludeValue(int groupId, float value)
+		void IncludeValue(int groupId, double value)
 		{
 			if(groupId == 0)
 			{
 				n1++;
 
-				float delta = value - m1;
+				double delta = value - m1;
 				m1 += delta / n1;
 
 				if(n1 > 1)
@@ -30,7 +30,7 @@ namespace BrainGraph { namespace Compute { namespace Graph
 			{
 				n2++;
 
-				float delta = value - m2;
+				double delta = value - m2;
 				m2 += delta / n2;
 
 				if(n2 > 1)
@@ -40,14 +40,14 @@ namespace BrainGraph { namespace Compute { namespace Graph
 
 		TStat Calculate()
 		{
-			float v1 = abs(dv1) / ( n1 - 1 );
-			float v2 = abs(dv2) / ( n2 - 1 );
+			double v1 = abs(dv1) / ( n1 - 1 );
+			double v2 = abs(dv2) / ( n2 - 1 );
 			
-			float tstat = 0;
+			double tstat = 0;
 			if(v1 < 0.00000001f && v2 < 0.00000001f)
 				tstat = 0;
 			else
-				tstat = (m1 - m2) / sqrt( ( v1 / (float)n1 ) + ( v2 / (float)n2 ) );
+				tstat = (m1 - m2) / sqrt( ( v1 / (double)n1 ) + ( v2 / (double)n2 ) );
 
 			TStat stat;
 			stat.V1 = v1;
@@ -61,7 +61,7 @@ namespace BrainGraph { namespace Compute { namespace Graph
 		
 	private:
 		int n1, n2;
-		float m1, m2;
-		float dv1, dv2;
+		double m1, m2;
+		double dv1, dv2;
 	};
 }}}
